style: format all files and add formatter

+++ scripts.nix
@@ -4,30 +4,36 @@ let
   hostName = if osConfig != null then osConfig.networking.hostName else "nixos-x390";
 in
 {
-    home.packages = [
-        (pkgs.writeShellScriptBin "nix-deploy" ''
-        set -e
+  home.packages = [
+    (pkgs.writeShellScriptBin "nix-deploy" ''
+      set -e
 
-        git add -A
+      if [ ! -f "flake.nix" ]; then
+          echo "error: there is no flake.nix" >&2
+          exit 1
+      fi
 
-        if [ "$1" = "--amend" ]; then
-            git commit --amend --no-edit || echo "No changes to amend"
+      nixpkgs-fmt .
 
-        elif [ "$2" = "--amend" ]; then
-            git commit --amend -m "$1" || echo "No changes to amend"
+      git add -A
 
-        else
-            COMMIT_MSG=''${1:-"deploy: $(date '+%Y-%m-%d %H:%M')"}
-            git commit -m "$COMMIT_MSG" || echo "No changes to commit"
-        fi
+      if [ "$1" = "--amend" ]; then
+          git commit --amend --no-edit || echo "No changes to amend"
 
-        sudo rsync -trq --chown=root:root \
-        --include={"configuration.nix","hardware-configuration.nix","flake.nix","flake.lock"} \
-        --exclude="*" \
-        /home/fkngoose/nixos-config/ /etc/nixos/
+      elif [ "$2" = "--amend" ]; then
+          git commit --amend -m "$1" || echo "No changes to amend"
 
-        sudo nixos-rebuild switch --flake .#${hostName}
-        '')
-    ];
+      else
+          COMMIT_MSG=''${1:-"deploy: $(date '+%Y-%m-%d %H:%M')"}
+          git commit -m "$COMMIT_MSG" || echo "No changes to commit"
+      fi
 
+      sudo rsync -trq --chown=root:root \
+      --include={"configuration.nix","hardware-configuration.nix","flake.nix","flake.lock"} \
+      --exclude="*" \
+      ./ /etc/nixos/
+
+      sudo nixos-rebuild switch --flake .#${hostName}
+    '')
+  ];
 }
